Patch 16.6.1

+++ src/ProfilePage/ProfilePageCustomer.jsx
@@ -210,12 +210,13 @@ const ProfilePageCustomer = () => {
             })
 
             try {
-                if (filterReceipt.length > 0 && filter.length > 0) {
-                    await deleteDoc(doc(db, 'Receipt', filterReceipt[0].docId));
-                    await deleteDoc(doc(db, 'Order', filter[0].docId))
-                    await deleteDoc(doc(db, 'Customer', auth?.currentUser?.uid))
-                    logout();
-                }
+
+                if (filterReceipt.length > 0) await deleteDoc(doc(db, 'Receipt', filterReceipt[0].docId));
+                if (filter.length > 0) await deleteDoc(doc(db, 'Order', filter[0].docId))
+
+                await deleteDoc(doc(db, 'Customer', auth?.currentUser?.uid))
+                logout();
+
             } catch (error) { console.error(error) }
